Checkup: Move affinity calculation to vmi
Currently, the affinity calculation is located under the `vmi`
package.

The above code is checkup-specific, and could be moved to the `checkup`
package, along with the other check-specific code.
